Dropped Maniac's M0.22 code over the current tree
Ok, first thing to note is that I didn't really review the code. All I
did was to convert the DOS line endings to UNIX and then look at the
diff output so it looks somewhat good.

Also some of the commits Igor did back in 2006 were overriden as this
diff is based on 2.8 release code and the CVS head was going towards 2.9
at the time of import.